Renaming ObjectPropertyComparator to PropertyComparisonWrapper
