+++ src/main/java/org/testmonkeys/jentitytest/Entity.java
@@ -4,7 +4,7 @@ import org.testmonkeys.jentitytest.matchers.EntityMatcher;
 
 public final class Entity {
 
-    public static EntityMatcher isEqualTo(Object expected) {
-        return new EntityMatcher(expected);
+    public static <T> EntityMatcher<T> isEqualTo(Object expected) {
+        return new EntityMatcher<>(expected);
     }
 }
