+++ src/main/java/org/testmonkeys/jentitytest/inspect/EntityInspector.java
@@ -1,9 +1,9 @@
 package org.testmonkeys.jentitytest.inspect;
 
 import org.testmonkeys.jentitytest.comparison.ComparisonModel;
+import org.testmonkeys.jentitytest.comparison.PropertyComparisonWrapper;
 import org.testmonkeys.jentitytest.comparison.property.SimpleTypeComparator;
 import org.testmonkeys.jentitytest.framework.JEntityTestException;
-import org.testmonkeys.jentitytest.matchers.ObjectPropertyComparator;
 
 import java.beans.BeanInfo;
 import java.beans.IntrospectionException;
@@ -32,7 +32,7 @@ public class EntityInspector {
                 for (Annotation annotation : field.getAnnotations()) {
                     if (annotationToComparator.mapsToComparator(annotation)) {
                         model.setComparisonPoint(propertyDescriptor,
-                                new ObjectPropertyComparator(annotationToComparator.getComparatorForAnnotation(annotation)));
+                                new PropertyComparisonWrapper(annotationToComparator.getComparatorForAnnotation(annotation)));
                         customComparison = true;
                         fieldLevelComparison = true;
                     }
@@ -44,13 +44,13 @@ public class EntityInspector {
                 for (Annotation annotation : method.getAnnotations()) {
                     if (annotationToComparator.mapsToComparator(annotation)) {
                         model.setComparisonPoint(propertyDescriptor,
-                                new ObjectPropertyComparator(annotationToComparator.getComparatorForAnnotation(annotation)));
+                                new PropertyComparisonWrapper(annotationToComparator.getComparatorForAnnotation(annotation)));
                         customComparison = true;
                     }
                 }
             }
             if (!customComparison)
-                model.setComparisonPoint(propertyDescriptor, new ObjectPropertyComparator(new SimpleTypeComparator()));
+                model.setComparisonPoint(propertyDescriptor, new PropertyComparisonWrapper(new SimpleTypeComparator()));
         }
         return model;
     }
